+++ cms/components/BlogPostPreview.jsx
@@ -1,0 +1,13 @@
+import React from 'react';
+
+import { BlogPostTemplate } from '../../src/templates/blog-post';
+
+const BlogPostPreview = ({ entry, widgetFor }) => (
+  <BlogPostTemplate
+    content={widgetFor('body')}
+    description={entry.getIn(['data', 'description'])}
+    title={entry.getIn(['data', 'title'])}
+  />
+);
+
+export default BlogPostPreview;
